+++ SistemaAcademico/Controllers/EvaluationController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using AcademicSystemApi.Extensions;
 using BLL.Interfaces;
 using Entities;
 using Microsoft.AspNetCore.Authorization;
@@ -18,16 +19,21 @@ namespace AcademicSystemApi.Controllers
     {
 
         IEvaluationService _service;
-        public EvaluationController(IEvaluationService service)
+        IUserService _userService;
+        IStudentService _studentService;
+        IInstructorService _instructorService;
+        public EvaluationController(IEvaluationService service, IUserService userService, IStudentService studentService, IInstructorService instructorService)
         {
             this._service = service;
+            this._userService = userService;
+            this._studentService = studentService;
+            this._instructorService = instructorService;
         }
 
+        /*
         [Authorize]
         public async Task<object> GetEvaluations()
         {
-
-
             try
             {
                 DataResponse<Evaluation> response = await _service.GetAll();
@@ -43,6 +49,33 @@ namespace AcademicSystemApi.Controllers
                 return null;
             }
         }
+        */
+
+        private async Task<bool> PermissionCheckToReadEvaluation(Evaluation e)
+        {
+            bool hasPermissionToRead = false;
+
+            User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+            if (user.Student != null)
+            {
+                Student student = (await this._studentService.GetByID(user.Student.ID)).Data[0];
+                // ver se a evaluation StudentID == Student.ID
+                if (student.Evaluations.Where(evaluation => evaluation.StudentID == e.StudentID).Any())
+                {
+                    hasPermissionToRead = true;
+                }
+            }
+            if (user.Instructor != null)
+            {
+                Instructor instructor = (await this._instructorService.GetByID(user.Instructor.ID)).Data[0];
+                if (instructor.Classes.Where(instructorClass => instructorClass.ClassID == e.ClassID).Any())
+                {
+                    hasPermissionToRead = true;
+                }
+            }
+
+            return hasPermissionToRead;
+        }
 
         [HttpGet]
         [Route("{id}")]
@@ -53,16 +86,22 @@ namespace AcademicSystemApi.Controllers
             {
                 DataResponse<Evaluation> response = await _service.GetByID(id);
 
-                //response.Data.ForEach(Evaluation => Evaluation.Course.Evaluations = null);
-
-                return new
+  
+                if (response.HasError())
                 {
-                    success = response.Success,
-                    data = response.Success ? response.Data : null
-                };
+                    return response;
+                }
+
+                if (await this.PermissionCheckToReadEvaluation(response.Data[0]))
+                {
+                    return this.SendResponse(response);
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -73,14 +112,24 @@ namespace AcademicSystemApi.Controllers
         {
             try
             {
-                Response response = await _service.Create(Evaluation);
-                return new
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+
+                if (user.Instructor != null)
                 {
-                    success = response.Success
-                };
+                    Instructor instructor = (await this._instructorService.GetByID(user.Instructor.ID)).Data[0];
+
+                    if (instructor.Classes.Where(c => c.ClassID == Evaluation.ClassID).Any())
+                    {
+                        Response response = await _service.Create(Evaluation);
+                        return this.SendResponse(response);
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -94,14 +143,24 @@ namespace AcademicSystemApi.Controllers
             Evaluation.ID = id;
             try
             {
-                Response response = await _service.Update(Evaluation);
-                return new
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+
+                if (user.Instructor != null)
                 {
-                    success = response.Success
-                };
+                    Instructor instructor = (await this._instructorService.GetByID(user.Instructor.ID)).Data[0];
+
+                    if (instructor.Classes.Where(c => c.ClassID == Evaluation.ClassID).Any())
+                    {
+                        Response response = await _service.Update(Evaluation);
+                        return this.SendResponse(response);
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -113,15 +172,24 @@ namespace AcademicSystemApi.Controllers
         {
             try
             {
-                Response response = await _service.Delete(id);
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
 
-                return new
+                if (user.Instructor != null)
                 {
-                    success = response.Success
-                };
+                    Instructor instructor = (await this._instructorService.GetByID(user.Instructor.ID)).Data[0];
+                    Evaluation Evaluation = (await this._service.GetByID(id)).Data[0];
+                    if (instructor.Classes.Where(c => c.ClassID == Evaluation.ClassID).Any())
+                    {
+                        Response response = await _service.Delete(id);
+                        return this.SendResponse(response);
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
