+++ SistemaAcademico/Controllers/SubjectController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using AcademicSystemApi.Extensions;
 using BLL.Interfaces;
 using Entities;
 using Microsoft.AspNetCore.Authorization;
@@ -18,11 +19,25 @@ namespace AcademicSystemApi.Controllers
     public class SubjectController : ControllerBase
     {
         ISubjectService _service;
-        public SubjectController(ISubjectService service)
+        IUserService _userService;
+        IInstructorService _instructorService;
+        IStudentService _studentService;
+        ICoordinatorService _coordinatorService;
+        IClassService _classService;
+        ICourseService _courseService;
+        IOwnerService _ownerService;
+        public SubjectController(ISubjectService service, IUserService userService, IInstructorService instructorService, IStudentService studentService, ICoordinatorService coordinatorService, IClassService classService, ICourseService courseService, IOwnerService ownerService)
         {
             this._service = service;
+            this._userService = userService;
+            this._instructorService = instructorService;
+            this._studentService = studentService;
+            this._coordinatorService = coordinatorService;
+            this._classService = classService;
+            this._courseService = courseService;
+            this._ownerService = ownerService;
         }
-
+        /*
         [Authorize]
         public async Task<object> GetSubjects()
         {
@@ -41,6 +56,54 @@ namespace AcademicSystemApi.Controllers
                 return null;
             }
         }
+        */
+        private async Task<bool> PermissionCheckToReadSubject(Subject subject)
+        {
+
+            User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+            if (user.Instructor != null)
+            {
+                Instructor instructor = (await this._instructorService.GetByID(user.Instructor.ID)).Data[0];
+                if (subject.Instructors != null)
+                {
+                    foreach (SubjectInstructor subjectInstructor in subject.Instructors)
+                    {
+                        if (instructor.Subjects.Where(s => s.InstructorID == subjectInstructor.InstructorID).Any()) 
+                        {
+                            return true;
+                        }
+                    }
+                }
+            }
+
+            if (user.Student != null)
+            {
+                Student student = (await this._studentService.GetByID(user.Student.ID)).Data[0];
+                foreach (Class Class in subject.Classes)
+                {
+                    if (student.Classes.Where(c => c.ClassID == Class.ID).Any())
+                    {
+                        return true;
+                    }
+                }
+            }
+
+            if (user.Coordinator != null)
+            {
+                Coordinator coordinator = (await this._coordinatorService.GetByID(user.Coordinator.ID)).Data[0];
+
+                foreach (CoordinatorClass coordinatorClass in coordinator.Classes)
+                {
+                    Class Class = (await this._classService.GetByID(coordinatorClass.ClassID)).Data[0];
+                    if (Class.CourseID == subject.CourseID)
+                    {
+                        return true;
+                    }
+                }
+            }
+
+            return false;
+        }
 
         [HttpGet]
         [Route("{id}")]
@@ -51,16 +114,21 @@ namespace AcademicSystemApi.Controllers
             {
                 DataResponse<Subject> response = await _service.GetByID(id);
 
-                response.Data.ForEach(subject => subject.Course.Subjects = null);
-
-                return new
+                if (response.HasError())
                 {
-                    success = response.Success,
-                    data = response.Success ? response.Data : null
-                };
+                    return response;
+                }
+
+                if (await this.PermissionCheckToReadSubject(response.Data[0]))
+                {
+                    return this.SendResponse(response);
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -71,14 +139,28 @@ namespace AcademicSystemApi.Controllers
         {
             try
             {
-                Response response = await _service.Create(Subject);
-                return new
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+
+                if (user.Owner != null)
                 {
-                    success = response.Success
-                };
+                    Owner owner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
+
+                    foreach (OwnerCourse oc in owner.Courses)
+                    {
+                        Course course = (await _courseService.GetByID(oc.CourseID)).Data[0];
+                        if (course.Classes.Where(c => c.ID == Subject.CourseID).Any())
+                        {
+                            Response response = await _service.Create(Subject);
+                            return this.SendResponse(response);
+                        }
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -91,14 +173,28 @@ namespace AcademicSystemApi.Controllers
             Subject.ID = id; 
             try
             {
-                Response response = await _service.Update(Subject);
-                return new
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
+
+                if (user.Owner != null)
                 {
-                    success = response.Success
-                };
+                    Owner owner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
+
+                    foreach (OwnerCourse oc in owner.Courses)
+                    {
+                        Course course = (await _courseService.GetByID(oc.CourseID)).Data[0];
+                        if (course.Classes.Where(c => c.ID == Subject.CourseID).Any())
+                        {
+                            Response response = await _service.Update(Subject);
+                            return this.SendResponse(response);
+                        }
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
@@ -110,15 +206,28 @@ namespace AcademicSystemApi.Controllers
         {
             try
             {
-                Response response = await _service.Delete(id);
+                User user = (await this._userService.GetByID(this.GetUserID())).Data[0];
 
-                return new
+                if (user.Owner != null)
                 {
-                    success = response.Success
-                };
+                    Owner owner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
+
+                    foreach (OwnerCourse oc in owner.Courses)
+                    {
+                        Course course = (await _courseService.GetByID(oc.CourseID)).Data[0];
+                        if (course.Classes.Where(c => c.ID == id).Any())
+                        {
+                            Response response = await _service.Delete(id);
+                            return this.SendResponse(response);
+                        }
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception e)
             {
+                Response.StatusCode = StatusCode(500).StatusCode;
                 return null;
             }
         }
